Move components around.
