+++ client/src/archived_components/Vulnerabilities.css
