+++ client/src/archived_components/Vulnerabilities.jsx
