+++ client/src/archived_components/VulnerabilitiesLoader.jsx
