+++ client/src/components/adaptations/AdaptationCategories.js
