+++ client/src/components/adaptations/StaticAdaptation.css
