+++ client/src/components/adaptations/StaticAdaptation.jsx
