+++ client/src/components/adaptations/StaticAdaptations.jsx
