+++ client/src/components/adaptations/StaticReferences.css
