+++ client/src/components/adaptations/StaticReferences.jsx
