+++ client/src/components/footer/FooterText.jsx
